feat: 리뷰 내용에 따른 아이템 동적 높이 구현
